Consolidate and document utils
