+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/util/PylonUtils.kt
@@ -2,45 +2,47 @@
 
 package io.github.pylonmc.pylon.core.util
 
-import com.destroystokyo.paper.profile.PlayerProfile
-import com.mojang.brigadier.context.CommandContext
+import io.github.pylonmc.pylon.core.PylonCore
 import io.github.pylonmc.pylon.core.addon.PylonAddon
 import io.github.pylonmc.pylon.core.entity.display.transform.TransformUtil.yawToCardinalDirection
-import io.github.pylonmc.pylon.core.fluid.PylonFluid
 import io.github.pylonmc.pylon.core.item.PylonItem
-import io.github.pylonmc.pylon.core.recipe.PylonRecipe
 import io.github.pylonmc.pylon.core.registry.PylonRegistry
-import io.papermc.paper.command.brigadier.CommandSourceStack
-import io.papermc.paper.command.brigadier.argument.resolvers.BlockPositionResolver
-import io.papermc.paper.command.brigadier.argument.resolvers.FinePositionResolver
-import io.papermc.paper.command.brigadier.argument.resolvers.PlayerProfileListResolver
-import io.papermc.paper.command.brigadier.argument.resolvers.RotationResolver
-import io.papermc.paper.command.brigadier.argument.resolvers.selector.EntitySelectorArgumentResolver
-import io.papermc.paper.command.brigadier.argument.resolvers.selector.PlayerSelectorArgumentResolver
-import io.papermc.paper.math.BlockPosition
-import io.papermc.paper.math.FinePosition
-import io.papermc.paper.math.Rotation
+import io.papermc.paper.datacomponent.DataComponentType
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.TranslatableComponent
 import net.kyori.adventure.text.TranslationArgumentLike
+import net.kyori.adventure.text.minimessage.MiniMessage
 import org.bukkit.NamespacedKey
 import org.bukkit.block.BlockFace
-import org.bukkit.entity.Entity
 import org.bukkit.entity.Player
 import org.bukkit.event.Event
+import org.bukkit.inventory.Inventory
+import org.bukkit.inventory.ItemStack
+import org.bukkit.persistence.PersistentDataContainer
+import org.bukkit.persistence.PersistentDataHolder
+import org.bukkit.persistence.PersistentDataType
 import org.bukkit.util.Vector
 import org.joml.RoundingMode
 import org.joml.Vector3f
 import org.joml.Vector3i
+import java.lang.invoke.MethodHandle
+import java.lang.invoke.MethodHandles
 import kotlin.math.absoluteValue
-import kotlin.reflect.typeOf
-
-fun NamespacedKey.isFromAddon(addon: PylonAddon): Boolean {
-    return namespace == addon.key.namespace
-}
+import kotlin.properties.ReadWriteProperty
+import kotlin.reflect.KProperty
 
 /**
- * Converts an approximate orthogonal vector to a [BlockFace]
+ * Checks whether a [NamespacedKey] is from [addon]
+ */
+@JvmName("isKeyFromAddon")
+fun NamespacedKey.isFromAddon(addon: PylonAddon): Boolean
+    = namespace == addon.key.namespace
+
+/**
+ * Converts an orthogonal vector to a [BlockFace]
+ *
+ *  @return The face that the vector is facing
+ *  @throws IllegalStateException if the vector is not pointing in a cardinal direction
  */
 fun vectorToBlockFace(vector: Vector3i): BlockFace {
     return if (vector.x > 0 && vector.y == 0 && vector.z == 0) {
@@ -60,13 +62,30 @@ fun vectorToBlockFace(vector: Vector3i): BlockFace {
     }
 }
 
+/**
+ * Converts an orthogonal vector to a [BlockFace]
+ *
+ *  @return The face that the vector is facing
+ *  @throws IllegalStateException if the vector is not pointing in a cardinal direction
+ */
 fun vectorToBlockFace(vector: Vector3f) = vectorToBlockFace(Vector3i(vector, RoundingMode.HALF_DOWN))
 
+/**
+ * Converts an orthogonal vector to a [BlockFace]
+ *
+ *  @return The face that the vector is facing
+ *  @throws IllegalStateException if the vector is not pointing in a cardinal direction
+ */
 // use toVector3f rather than toVector3i because toVector3i will floor components
 fun vectorToBlockFace(vector: Vector) = vectorToBlockFace(vector.toVector3f())
 
 /**
- * Rotates a BlockFace to the player's reference frame. Where the player is facing becomes NORTH.
+ * Rotates a BlockFace to the [player]'s reference frame.
+ *
+ * @param player The player to act as the reference frame. Where the player is facing becomes NORTH.
+ * @param face The face to rotate
+ * @param allowVertical Whether we should include UP and DOWN
+ * @return The block face rotated to the player's reference frame
  */
 fun rotateToPlayerFacing(player: Player, face: BlockFace, allowVertical: Boolean): BlockFace {
     var vector = face.direction.clone().rotateAroundY(yawToCardinalDirection(player.eyeLocation.yaw).toDouble())
@@ -79,25 +98,13 @@ fun rotateToPlayerFacing(player: Player, face: BlockFace, allowVertical: Boolean
     return vectorToBlockFace(vector)
 }
 
-fun isCardinalDirection(vector: Vector3i) = (vector.x != 0 && vector.y == 0 && vector.z == 0)
-        || (vector.x == 0 && vector.y != 0 && vector.z == 0)
-        || (vector.x == 0 && vector.y == 0 && vector.z != 0)
-
-fun isCardinalDirection(vector: Vector3f)
-    = (vector.x.absoluteValue > 1.0e-6 && vector.y.absoluteValue < 1.0e-6 && vector.z.absoluteValue < 1.0e-6)
-        || (vector.x.absoluteValue < 1.0e-6 && vector.y.absoluteValue > 1.0e-6 && vector.z.absoluteValue < 1.0e-6)
-        || (vector.x.absoluteValue < 1.0e-6 && vector.y.absoluteValue < 1.0e-6 && vector.z.absoluteValue > 1.0e-6)
-
-fun getAddon(key: NamespacedKey): PylonAddon =
-    PylonRegistry.Companion.ADDONS.find { addon -> addon.key.namespace == key.namespace }
-        ?: error("Key does not have a corresponding addon; does your addon call registerWithPylon()?")
-
 /**
- * Rotates a vector to face a direction
+ * Rotates [vector] to face a direction
  *
- * The direction given must be a horizontal cardinal direction (north, east, south, west)
+ * Assumes north to be the default direction (i.e. supplying north will result in no rotation)
  *
- * Assumes north to be the default direction (supplying north will result in no rotation)
+ * @param face Must be a horizontal cardinal direction (north, east, south, west)
+ * @return The rotated vector
  */
 fun rotateVectorToFace(vector: Vector3i, face: BlockFace) = when (face) {
     BlockFace.NORTH -> vector
@@ -107,47 +114,33 @@ fun rotateVectorToFace(vector: Vector3i, face: BlockFace) = when (face) {
     else -> throw IllegalArgumentException("$face is not a horizontal cardinal direction")
 }
 
-fun wrapText(text: String, limit: Int): List<String> {
-    val words = text.split(" ")
-    val lines = mutableListOf<String>()
-    var currentLine = StringBuilder()
+/**
+ * @return Whether [vector] is a cardinal direction
+ */
+fun isCardinalDirection(vector: Vector3i) = (vector.x != 0 && vector.y == 0 && vector.z == 0)
+        || (vector.x == 0 && vector.y != 0 && vector.z == 0)
+        || (vector.x == 0 && vector.y == 0 && vector.z != 0)
 
-    for (word in words) {
-        if (currentLine.length + word.length + 1 > limit) {
-            currentLine.append(' ')
-            lines.add(currentLine.toString())
-            currentLine = StringBuilder()
-        }
-        if (currentLine.isNotEmpty()) {
-            currentLine.append(" ")
-        }
-        currentLine.append(word)
-    }
-    if (currentLine.isNotEmpty()) {
-        lines.add(currentLine.toString())
-    }
-    return lines
-}
+/**
+ * @return Whether [vector] is a cardinal direction
+ */
+fun isCardinalDirection(vector: Vector3f)
+    = (vector.x.absoluteValue > 1.0e-6 && vector.y.absoluteValue < 1.0e-6 && vector.z.absoluteValue < 1.0e-6)
+        || (vector.x.absoluteValue < 1.0e-6 && vector.y.absoluteValue > 1.0e-6 && vector.z.absoluteValue < 1.0e-6)
+        || (vector.x.absoluteValue < 1.0e-6 && vector.y.absoluteValue < 1.0e-6 && vector.z.absoluteValue > 1.0e-6)
 
-@JvmSynthetic
-@Suppress("UnstableApiUsage")
-inline fun <reified T> CommandContext<CommandSourceStack>.getArgument(name: String): T {
-    return when (typeOf<T>()) {
-        typeOf<BlockPosition>() -> getArgument(name, BlockPositionResolver::class.java).resolve(source)
-        typeOf<List<Entity>>() -> getArgument(name, EntitySelectorArgumentResolver::class.java).resolve(source)
-        typeOf<Entity>() -> getArgument(name, EntitySelectorArgumentResolver::class.java).resolve(source).first()
-        typeOf<FinePosition>() -> getArgument(name, FinePositionResolver::class.java).resolve(source)
-        typeOf<List<PlayerProfile>>() -> getArgument(name, PlayerProfileListResolver::class.java).resolve(source)
-        typeOf<PlayerProfile>() -> getArgument(name, PlayerProfileListResolver::class.java).resolve(source).first()
-        typeOf<List<Player>>() -> getArgument(name, PlayerSelectorArgumentResolver::class.java).resolve(source)
-        typeOf<Player>() -> getArgument(name, PlayerSelectorArgumentResolver::class.java).resolve(source).first()
-        typeOf<Rotation>() -> getArgument(name, RotationResolver::class.java).resolve(source)
-        else -> getArgument(name, T::class.java)
-    } as T
-}
+/**
+ * @return The addon that [key] belongs to
+ */
+fun getAddon(key: NamespacedKey): PylonAddon =
+    PylonRegistry.Companion.ADDONS.find { addon -> addon.key.namespace == key.namespace }
+        ?: error("Key does not have a corresponding addon; does your addon call registerWithPylon()?")
 
 /**
  * Attaches arguments to a component and all its children.
+ *
+ * @param args List of arguments to attach
+ * @return The component with the arguments attached
  */
 @JvmName("attachArguments")
 fun Component.withArguments(args: List<TranslationArgumentLike>): Component {
@@ -159,64 +152,186 @@ fun Component.withArguments(args: List<TranslationArgumentLike>): Component {
     return result.children(result.children().map { it.withArguments(args) })
 }
 
-fun findRecipeFor(item: PylonItem): PylonRecipe? {
-    // 1. if there's a recipe with the same key as the item, use that
-    PylonRegistry.RECIPE_TYPES
-        .map { it.getRecipe(item.schema.key) }
-        .find { it != null }?.let { return it }
-
-    // 2. if there's a recipe which produces *only* that item, use that
-    // 3. if there's multiple recipes which produce only that item, choose the *lowest* one lexographically
-    val singleOutputRecipes = PylonRegistry.RECIPE_TYPES.asSequence()
-        .flatMap { it.recipes }
-        .filter { recipe -> recipe.isOutput(item.stack) && recipe.results.size == 1 }
-        .sortedBy { it.key }
-        .toList()
-
-    if (singleOutputRecipes.isNotEmpty()) return singleOutputRecipes.first()
-
-    // 4. if there's a recipe which produces the item *alongside* other things, use that
-    // 5. if there's multiple recipes which produce the item alongside other things, choose the *lowest* one lexographically
-    val multiOutputRecipes = PylonRegistry.RECIPE_TYPES.asSequence()
-        .flatMap { it.recipes }
-        .filter { recipe -> recipe.isOutput(item.stack) }
-        .sortedBy { it.key }
-        .toList()
-
-    if (multiOutputRecipes.isNotEmpty()) return multiOutputRecipes.first()
-
-    return null
-}
-
-fun findRecipeFor(fluid: PylonFluid): PylonRecipe? {
-    // 1. if there's a recipe with the same key as the item, use that
-    PylonRegistry.RECIPE_TYPES
-        .map { it.getRecipe(fluid.key) }
-        .find { it != null }?.let { return it }
-
-    // 2. if there's a recipe which produces *only* that item, use that
-    // 3. if there's multiple recipes which produce only that item, choose the *lowest* one lexographically
-    val singleOutputRecipes = PylonRegistry.RECIPE_TYPES.asSequence()
-        .flatMap { it.recipes }
-        .filter { recipe -> recipe.isOutput(fluid) && recipe.results.size == 1 }
-        .sortedBy { it.key }
-        .toList()
-
-    if (singleOutputRecipes.isNotEmpty()) return singleOutputRecipes.first()
-
-    // 4. if there's a recipe which produces the item *alongside* other things, use that
-    // 5. if there's multiple recipes which produce the item alongside other things, choose the *lowest* one lexographically
-    val multiOutputRecipes = PylonRegistry.RECIPE_TYPES.asSequence()
-        .flatMap { it.recipes }
-        .filter { recipe -> recipe.isOutput(fluid) }
-        .sortedWith { a: PylonRecipe, b: PylonRecipe -> a.key.compareTo(b.key) }
-        .toList()
-
-    if (multiOutputRecipes.isNotEmpty()) return multiOutputRecipes.first()
-
-    return null
-}
-
+/**
+ * (Heuristically) checks whether an event is 'fake' (by checking if it has 'Fake' in its name)
+ *
+ * 'Fake' events are often used to check actions before performing them.
+ */
 fun isFakeEvent(event: Event): Boolean {
     return event.javaClass.name.contains("Fake")
 }
+
+/**
+ * [BlockFace.UP], [BlockFace.DOWN], [BlockFace.EAST], [BlockFace.WEST], [BlockFace.SOUTH], [BlockFace.NORTH]
+ */
+@JvmField
+val IMMEDIATE_FACES: Array<BlockFace> = arrayOf(
+    BlockFace.UP,
+    BlockFace.DOWN,
+    BlockFace.EAST,
+    BlockFace.WEST,
+    BlockFace.SOUTH,
+    BlockFace.NORTH
+)
+
+/**
+ * Same as [IMMEDIATE_FACES] but includes diagonal faces, not including the vertical directions.
+ */
+@JvmField
+val IMMEDIATE_FACES_WITH_DIAGONALS: Array<BlockFace> = arrayOf(
+    BlockFace.UP,
+    BlockFace.DOWN,
+    BlockFace.EAST,
+    BlockFace.WEST,
+    BlockFace.SOUTH,
+    BlockFace.NORTH,
+    BlockFace.NORTH_EAST,
+    BlockFace.NORTH_WEST,
+    BlockFace.SOUTH_EAST,
+    BlockFace.SOUTH_WEST,
+    BlockFace.EAST
+)
+
+@JvmSynthetic
+internal fun pylonKey(key: String): NamespacedKey = NamespacedKey(PylonCore, key)
+
+@JvmSynthetic
+internal fun Class<*>.findConstructorMatching(vararg types: Class<*>): MethodHandle? {
+    return declaredConstructors.firstOrNull {
+        it.parameterTypes.size == types.size &&
+                it.parameterTypes.zip(types).all { (param, given) -> given.isSubclassOf(param) }
+    }?.let(MethodHandles.lookup()::unreflectConstructor)
+}
+
+// I can never remember which way around `isAssignableFrom` goes,
+// so this is a helper function to make it more readable
+@JvmSynthetic
+private fun Class<*>.isSubclassOf(other: Class<*>): Boolean = other.isAssignableFrom(this)
+
+/**
+ * Small helper function to convert a minimessage string (eg: '<red>bruh') into a component
+ * @param string The string to turn into a component
+ * @returns The string as a component
+ */
+@JvmSynthetic
+fun fromMiniMessage(string: String): Component = MiniMessage.miniMessage().deserialize(string)
+
+/**
+ * Finds a Pylon item in an inventory. Use this to find Pylon items instead of traditional
+ * find methods, because this will compare Pylon IDs.
+ *
+ * @param inventory The inventory to search
+ * @param targetItem The item to find. Items will be compared by their Pylon ID
+ * @return The slot containing the item, or null if no item was found
+ */
+fun findPylonItemInInventory(inventory: Inventory, targetItem: PylonItem): Int? {
+    for (i in 0..<inventory.size) {
+        val item = inventory.getItem(i)?.let {
+            PylonItem.fromStack(it)
+        }
+        if (item == targetItem) {
+            return i
+        }
+    }
+    return null
+}
+
+/**
+ * Compares two Pylon items to check if they have the same Pylon ID. If
+ * neither item is a Pylon item, the material will be compared instead.
+ *
+ * @return Whether the items are the same.
+ */
+fun ItemStack?.isPylonSimilar(item2: ItemStack?): Boolean {
+    // Both items null
+    if (this == null && item2 == null) {
+        return true
+    }
+
+    // One item null, one not null
+    if (!(this != null && item2 != null)) {
+        return false
+    }
+
+    val pylonItem1 = PylonItem.fromStack(this)
+    val pylonItem2 = PylonItem.fromStack(item2)
+
+    // Both pylon items null
+    if (pylonItem1 == null && pylonItem2 == null) {
+        return this.isSimilar(item2)
+    }
+
+    // One pylon item null, one not null
+    if (!(pylonItem1 != null && pylonItem2 != null)) {
+        return false
+    }
+
+    return pylonItem1.schema.key == pylonItem2.schema.key
+}
+
+@JvmSynthetic
+inline fun <reified T> ItemStack?.isPylonAndIsNot(): Boolean {
+    val pylonItem = PylonItem.fromStack(this)
+    return pylonItem != null && pylonItem !is T
+}
+
+@JvmSynthetic
+@Suppress("UnstableApiUsage")
+inline fun <T : Any> ItemStack.editData(
+    type: DataComponentType.Valued<T>,
+    block: (T) -> T
+): ItemStack {
+    val data = getData(type) ?: return this
+    setData(type, block(data))
+    return this
+}
+
+/**
+ * Wrapper around [PersistentDataContainer.set] that allows nullable values to be passed
+ *
+ * @param value The value to set. If this is null, the key will be removed from the container
+ */
+fun <P, C> PersistentDataContainer.setNullable(key: NamespacedKey, type: PersistentDataType<P, C>, value: C?) {
+    if (value != null) {
+        set(key, type, value)
+    } else {
+        remove(key)
+    }
+}
+
+/**
+ * Acts as a property delegate for stuff contained inside a [PersistentDataContainer]
+ * For example:
+ * ```
+ * val numberOfTimesJumped: Int by persistentData(NamespacedKey(yourPlugin, "jumped"), PersistentDataType.INTEGER) { 0 }
+ * ```
+ */
+@JvmSynthetic
+inline fun <T> persistentData(
+    key: NamespacedKey,
+    type: PersistentDataType<*, T & Any>,
+    crossinline default: () -> T
+) = object : ReadWriteProperty<PersistentDataHolder, T> {
+
+    override fun getValue(thisRef: PersistentDataHolder, property: KProperty<*>): T {
+        return thisRef.persistentDataContainer.get(key, type) ?: default()
+    }
+
+    override fun setValue(thisRef: PersistentDataHolder, property: KProperty<*>, value: T) {
+        if (value == null) {
+            thisRef.persistentDataContainer.remove(key)
+        } else {
+            thisRef.persistentDataContainer.set(key, type, value)
+        }
+    }
+}
+
+/**
+ * Same as [persistentData] but with a default value that is constant
+ */
+@JvmSynthetic
+fun <T> persistentData(
+    key: NamespacedKey,
+    type: PersistentDataType<*, T & Any>,
+    default: T
+) = persistentData(key, type) { default }
