+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/util/gui/ProgressItem.kt
@@ -14,11 +14,19 @@ import xyz.xenondevs.invui.item.ItemProvider
 import xyz.xenondevs.invui.item.impl.AbstractItem
 import java.time.Duration
 
+/**
+ * An item that counts down (by default) from a maximum durability, symbolising progress.
+ *
+ * For example, this might be used in a custom furnace to show how much burn time is left on
+ * the current fuel. In this case, you'd want override [totalTime] and set it to the the total
+ * burn time of the fuel (eg: 80 seconds for coal), and then gradually increase [progress] as the
+ * remaining burn time decreased.
+ *
+ * @param material The material to use for the item
+ * @param inverse If true, the progress bar will be inverted, meaning that 0.0 is full and 1.0 is empty.
+ */
 abstract class ProgressItem @JvmOverloads constructor(
     private val material: Material,
-    /**
-     * If true, the progress bar will be inverted, meaning that 0.0 is full and 1.0 is empty.
-     */
     private val inverse: Boolean = false
 ) : AbstractItem() {
 
@@ -57,6 +65,11 @@ abstract class ProgressItem @JvmOverloads constructor(
         return builder
     }
 
+    /**
+     * This can be used to modify the item being displayed.
+     *
+     * @param builder The item builder representing the progress item
+     */
     protected abstract fun completeItem(builder: ItemStackBuilder)
 
     override fun handleClick(clickType: ClickType, player: Player, event: InventoryClickEvent) {}
